third_party/blink/web_tests/http/tests/security/local-video-poster-from-remote-expected.txt

CONSOLE ERROR: Not allowed to load local resource: compass.jpg
Test that a remote video element can not use a local poster

This test requires the run-webkit httpd server (run-webkit-httpd)

SUCCESS: Local image failed to load from remote page.
